+++ scripts/__init__.py
@@ -1,2 +1,4 @@
 from .extractAllEditions import main as editionExtraction
 from .updateOldInTechOpenLinks import main as updateInTechOpenLinks
+from .updateLinksCatalog import main as updateLinks
+from .updateRecordsCatalogLink import main as updateRecordCatalogs